added test_set_personal_manager_rating_without_pm test and make code refactor
