+++ base/seleniumbase.py
@@ -25,19 +25,24 @@ class SeleniumBase:
         return locating[find_by]
 
     def is_visible(self, find_by: str, locator: str, locator_name: str = None) -> WebElement:
-        return self.__wait.until(expected.visibility_of_element_located((self.__get_selenium_by(find_by), locator)), locator_name)
+        return self.__wait.until(expected.visibility_of_element_located((self.__get_selenium_by(find_by), locator)),
+                                 locator_name)
 
     def is_present(self, find_by: str, locator: str, locator_name: str = None) -> WebElement:
-        return self.__wait.until(expected.presence_of_element_located((self.__get_selenium_by(find_by), locator)), locator_name)
+        return self.__wait.until(expected.presence_of_element_located((self.__get_selenium_by(find_by), locator)),
+                                 locator_name)
 
     def is_not_present(self, find_by: str, locator: str, locator_name: str = None) -> WebElement:
-        return self.__wait.until(expected.invisibility_of_element_located((self.__get_selenium_by(find_by), locator)), locator_name)
+        return self.__wait.until(expected.invisibility_of_element_located((self.__get_selenium_by(find_by), locator)),
+                                 locator_name)
 
     def are_visible(self, find_by: str, locator: str, locator_name: str = None) -> List[WebElement]:
-        return self.__wait.until(expected.visibility_of_all_elements_located((self.__get_selenium_by(find_by), locator)), locator_name)
+        return self.__wait.until(
+            expected.visibility_of_all_elements_located((self.__get_selenium_by(find_by), locator)), locator_name)
 
     def are_present(self, find_by: str, locator: str, locator_name: str = None) -> List[WebElement]:
-        return self.__wait.until(expected.presence_of_all_elements_located((self.__get_selenium_by(find_by), locator)), locator_name)
+        return self.__wait.until(expected.presence_of_all_elements_located((self.__get_selenium_by(find_by), locator)),
+                                 locator_name)
 
     def get_text_webelements(self, elements: List[WebElement]) -> List[str]:
         return [element.text for element in elements]
