--- base/utils.py
@@ -1,7 +1,0 @@
-from typing import List
-
-
-class Utils:
-
-    def join_strings(self, str_list: List[str]) -> str:
-        pass
