+++ suite_tests/API/test_authorization.py
@@ -1,5 +1,6 @@
 import requests
 import unittest
+
 from base.apibase import Apibase
 
 
@@ -12,6 +13,7 @@ class authorization(unittest.TestCase):
                    }
         body = {"login": Apibase.get_login_credentials(self), "password": Apibase.get_password_credentials(self)}
         response = requests.post(Apibase.get_route_auth('login'), headers=Headers, json=body)
+
         assert response.status_code == 200, 'User is authorized'
