+++ suite_tests/UI/test_logInForm.py
@@ -1,6 +1,7 @@
 import pytest
 from suite_tests.pom.login_pom import Login_POM
 
+
 @pytest.mark.usefixtures('setup')
 class TestLogInForm:
